+++ pkg/types/userconfig/api.go
@@ -55,7 +55,7 @@ type Predictor struct {
 	Config                 map[string]interface{} `json:"config" yaml:"config"`
 	Env                    map[string]string      `json:"env" yaml:"env"`
 	SignatureKey           *string                `json:"signature_key" yaml:"signature_key"`
-	BatchSize              *int64                 `json:"batch_size" yaml:"batch_size"`
+	BatchSize              *int32                 `json:"batch_size" yaml:"batch_size"`
 
 	// measured in seconds
 	BatchTimeout *float64 `json:"batch_timeout" yaml:"batch_timeout"`
